storage: add onSet and onRemove signals
With the addition of these signals, every method depending on them has to be @system.
This meant an big code refactoring in almost every function in VECS. Safety and purity in
VECS didn't and will not change, only the method declaration will be @system if using or
depending on Signal.

Signed-off-by: João Lourenço <jlourenco5691@gmail.com>