entity: fix unused alias this
The opEquals method already gives the functionality wanted with the used of
alias signature this. This was an old implementation for Entity comparison. The
other use was to give the user the illusion that an Entity is just a number,
however the need for that is not existent, and as such this should be removed.

Signed-off-by: João Lourenço <jlourenco5691@gmail.com>