entitybuilder: refactor entities to entity and fix dependent unit tests
An entity builder should be about building an rntity and ease the syntax for any
operation involving an entity. Before it was restricted to only generate new
entities with or without components. Although it eases much of the repeated
syntax when creating multiple entities it lacks in the capacity to support other
features sharing the same easiness. The new sytax does not take away much of the
above and has the potential to wrap every call involving an entity.

Signed-off-by: João Lourenço <jlourenco5691@gmail.com>