entitymanager: refactor EntityManager to receive a delegate
The empty delegate will serve as a gateway for @safe pure nohrow @nogc code with
VECS. EntityManager depends on Signal, and Signal cannot force the user to
connect slots that have one or more of these attributes. The only way is to use
a common type to all 'void delegate(...)'. The issue with this, is that this type
is @system, making all the internal code that depends on it @system as well. By
giving the user the choice to restrict his slot types and manage his own safety
attributes, VECS achieves the best of both worlds. It can be used with @safe code
but does not force the user to it.

Signed-off-by: João Lourenço <jlourenco5691@gmail.com>